+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - AttachedGizmo base class that attaches a gizmo to another object
  - ComponentGizmo base class that automatically attaches a gizmo to another component
  - BehaviourGizmo base class that draws its gizmo conditionally based on the enabled state of an attached behaviour
+ - TriggerGizmo class that tracks the colliders entering and exiting a trigger